+++ OpenSim/Simulation/Model/MovingPathPoint.cpp
@@ -25,8 +25,10 @@
 // INCLUDES
 //=============================================================================
 #include "MovingPathPoint.h"
+#include <OpenSim/Common/Function.h>
 #include <OpenSim/Common/Constant.h>
 #include <OpenSim/Common/MultiplierFunction.h>
+#include <OpenSim/Simulation/Model/PhysicalFrame.h>
 #include <OpenSim/Simulation/SimbodyEngine/Coordinate.h>
 
 
@@ -44,7 +46,7 @@ using SimTK::Vec3;
 /*
  * Default constructor.
  */
-MovingPathPoint::MovingPathPoint() : PathPoint()
+MovingPathPoint::MovingPathPoint() : Super()
 {
     constructProperties();
 }
@@ -136,14 +138,6 @@ void MovingPathPoint::extendConnectToModel(Model& model)
         "MovingPathPoint:: Components of the path point location "
         "must depend on the same Coordinate. Condition: "
         "x_coordinate == y_coordinate == z_coordinate  FAILED.");
-
-    // Overwrite any setting of a MovingPathPoint's location property.
-    // MovingPathPoint should NOT derive from Station.
-    // TODO: We should not have Path specific Points of any kind. They should
-    // be "path" Points by virtue of being subcomponents of a GeometryPath.
-    // Set it to NaN so it blows up if we attempt to use a MovingPathPoint
-    // as a Station.
-    set_location(SimTK::Vec3(SimTK::NaN));
 }
 
 //_____________________________________________________________________________
@@ -331,8 +325,6 @@ void MovingPathPoint::scale(const SimTK::Vec3& aScaleFactors)
             set_z_location(MultiplierFunction(get_z_location().clone(), aScaleFactors[2]));
         }
     }
-
-    updateGeometry();
 }