+++ OpenSim/Simulation/RegisterTypes_osimSimulation.cpp
@@ -84,7 +84,6 @@
 #include "Control/ControlConstant.h"
 #include "Control/ControlLinear.h"
 #include "Control/PrescribedController.h"
-#include "Control/ToyReflexController.h"
 
 #include "Wrap/PathWrap.h"
 #include "Wrap/PathWrapSet.h"
@@ -244,7 +243,6 @@ OSIMSIMULATION_API void RegisterTypes_osimSimulation()
 
     Object::registerType( ControlSetController() );
     Object::registerType( PrescribedController() );
-    Object::registerType( ToyReflexController() );
 
     Object::registerType( PathActuator() );
     Object::registerType( ProbeSet() );